move test data to __fixtures__
